requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
asyncpg
pydantic
pydantic-settings
python-dotenv
python-multipart
langgraph
langgraph-checkpoint-postgres
langchain
langchain-core
langchain-aws
langchain-openai
pgvector
httpx
orjson
pytest
pytest-asyncio
email-validator
psycopg2
pypdf
python-docx
